+++ src/lib/logical_query_plan/abstract_non_query_node.cpp
@@ -1,4 +1,4 @@
-#include "base_non_query_node.hpp"
+#include "abstract_non_query_node.hpp"
 
 #include "utils/assert.hpp"
