+++ src/lib/logical_query_plan/change_meta_table_node.hpp
@@ -4,7 +4,7 @@
 #include <string>
 #include <vector>
 
-#include "base_non_query_node.hpp"
+#include "abstract_non_query_node.hpp"
 
 namespace opossum {
