+++ src/lib/logical_query_plan/drop_table_node.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "base_non_query_node.hpp"
+#include "abstract_non_query_node.hpp"
 #include "enable_make_for_lqp_node.hpp"
 
 namespace opossum {
