+++ src/lib/logical_query_plan/drop_view_node.hpp
@@ -3,7 +3,7 @@
 #include <memory>
 #include <string>
 
-#include "base_non_query_node.hpp"
+#include "abstract_non_query_node.hpp"
 
 namespace opossum {
